Renaming exo4jax -> jaxoplanet (#10)
* renaming

* supporting merge queue
