+++ .github/workflows/tests.yml
@@ -7,6 +7,7 @@ on:
     tags:
       - "*"
   pull_request:
+  merge_group:
 
 jobs:
   tests:
