+++ pyproject.toml
@@ -14,13 +14,13 @@ include_trailing_comma = true
 force_grid_wrap = 0
 use_parentheses = true
 combine_as_imports = true
-known_first_party = ["exo4jax"]
+known_first_party = ["jaxoplanet"]
 
 [tool.coverage.run]
 parallel = true
 branch = true
-source = ["exo4jax"]
-omit = ["*_test.py", "*__init__*", "*/exo4jax/exo4jax.py"]
+source = ["jaxoplanet"]
+omit = ["*_test.py", "*__init__*", "*/jaxoplanet/jaxoplanet.py"]
 
 [tool.coverage.paths]
 source = ["src", "*/site-packages"]
