--- src/exo4jax/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from exo4jax import core as core, orbits as orbits
-from exo4jax.exo4jax_version import __version__ as __version__
-
-__uri__ = "https://github.com/exoplanet-dev/exo4jax"
-__author__ = "Dan Foreman-Mackey"
-__email__ = "foreman.mackey@gmail.com"
-__description__ = "Exoplanet, but this time in JAX"
-__license__ = "MIT"
