--- src/exo4jax/_src/experimental/starry/__init__.py
@@ -1,3 +1,0 @@
-from exo4jax._src.experimental.starry.solution import (
-    solution_vector as solution_vector,
-)
