--- src/exo4jax/core.py
@@ -1,2 +1,0 @@
-from exo4jax._src.core.kepler import kepler as kepler
-from exo4jax._src.core.quad import light_curve as light_curve
