--- src/exo4jax/light_curves.py
@@ -1,1 +1,0 @@
-from exo4jax._src.light_curves import QuadLightCurve as QuadLightCurve
