+++ src/jaxoplanet/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+from jaxoplanet import core as core, orbits as orbits
+from jaxoplanet.jaxoplanet_version import __version__ as __version__
+
+__uri__ = "https://github.com/exoplanet-dev/jaxoplanet"
+__author__ = "Dan Foreman-Mackey"
+__email__ = "foreman.mackey@gmail.com"
+__description__ = "Exoplanet, but this time in JAX"
+__license__ = "MIT"
