+++ src/jaxoplanet/_src/experimental/starry/__init__.py
@@ -1,0 +1,3 @@
+from jaxoplanet._src.experimental.starry.solution import (
+    solution_vector as solution_vector,
+)
