+++ src/jaxoplanet/light_curves.py
@@ -1,0 +1,1 @@
+from jaxoplanet._src.light_curves import QuadLightCurve as QuadLightCurve
