+++ src/jaxoplanet/orbits.py
@@ -1,6 +1,6 @@
-from exo4jax._src.keplerian import (
+from jaxoplanet._src.keplerian import (
     KeplerianBody as KeplerianBody,
     KeplerianCentral as KeplerianCentral,
     KeplerianOrbit as KeplerianOrbit,
 )
-from exo4jax._src.transit import TransitOrbit as TransitOrbit
+from jaxoplanet._src.transit import TransitOrbit as TransitOrbit
