+++ tests/core/limb_dark_test.py
@@ -8,8 +8,8 @@ import numpy as np
 import pytest
 from jax.test_util import check_grads
 
-from exo4jax._src.core.limb_dark import light_curve
-from exo4jax.test_utils import assert_allclose
+from jaxoplanet._src.core.limb_dark import light_curve
+from jaxoplanet.test_utils import assert_allclose
 
 
 @pytest.mark.parametrize("r", [0.01, 0.1, 1.1, 2.0])
