+++ tests/core/quad_test.py
@@ -4,8 +4,8 @@ import jax
 import jax.numpy as jnp
 import pytest
 
-from exo4jax._src.core.quad import quad_soln_impl
-from exo4jax.test_utils import assert_allclose
+from jaxoplanet._src.core.quad import quad_soln_impl
+from jaxoplanet.test_utils import assert_allclose
 
 
 @pytest.mark.skipif(
