+++ tests/experimental/starry/solution_test.py
@@ -6,8 +6,11 @@ import jax
 import numpy as np
 import pytest
 
-from exo4jax.test_utils import assert_allclose
-from exo4jax._src.experimental.starry.solution import kappas, solution_vector
+from jaxoplanet.test_utils import assert_allclose
+from jaxoplanet._src.experimental.starry.solution import (
+    kappas,
+    solution_vector,
+)
 
 
 def test_kappas():
